migrate create conversation modal
